homa-bay-laundry-hub/server/requirements.txt
Flask
Flask-JWT-Extended
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
requests 